src/flask_gsplat_demo/requirements.txt
flask==3.0.3
numpy==1.26.4
pillow==10.4.0
gsplat
tifffile
